feat(profile): refactor code .
